Oh my god it works I think. Model is never supposed to hold anything that will be shown. It does the calculations for ViewModel.  One can think of ViewModel as in charge of the situation. Model is the calculator. View Model is the one who types it in and shows the answer to View.
So its like this. We have our judges. The men who update the score board. then the score board itself.
When the score boards needs to be set, it is missing some numbers which the men who update the score board are supposed to provide.
The men of the score board though never do the math themselves. They go back stage and hand the judges enough information to calcuate a score, which the men who update the score board use to update it.